ocr_server/requirements.txt
fastapi
numpy
opencv-python-headless
pillow
pytesseract
python-multipart
uvicorn